test classes moved to correct module
