+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/main/java/com/swirlds/platform/test/SimpleEventGenerator.java
@@ -19,7 +19,7 @@ package com.swirlds.platform.test;
 import com.swirlds.common.system.NodeId;
 import com.swirlds.common.system.address.AddressBook;
 import com.swirlds.platform.internal.EventImpl;
-import com.swirlds.platform.test.event.RandomEventUtils;
+import com.swirlds.platform.test.fixtures.event.RandomEventUtils;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.HashMap;
 import java.util.HashSet;
