+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/main/java/com/swirlds/platform/test/consensus/NodeContext.java
@@ -27,9 +27,9 @@ import com.swirlds.platform.internal.EventImpl;
 import com.swirlds.platform.state.signed.SignedState;
 import com.swirlds.platform.test.event.emitter.CollectingEventEmitter;
 import com.swirlds.platform.test.event.emitter.EventEmitter;
-import com.swirlds.platform.test.event.source.EventSource;
 import com.swirlds.platform.test.fixtures.event.EventUtils;
 import com.swirlds.platform.test.fixtures.event.IndexedEvent;
+import com.swirlds.platform.test.fixtures.event.source.EventSource;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.Arrays;
