+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/main/java/com/swirlds/platform/test/event/source/ForkingEventSource.java
@@ -18,6 +18,7 @@ package com.swirlds.platform.test.event.source;
 
 import com.swirlds.common.test.fixtures.TransactionGenerator;
 import com.swirlds.platform.test.fixtures.event.IndexedEvent;
+import com.swirlds.platform.test.fixtures.event.source.AbstractEventSource;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.Random;
